Enumerate a new "password v2" protocol.
This adds the constants and some headers necessary to implement a
second version of the password protocol.  It uses a slight twist
on the AKEP2 algorithm found in the password v1 protocol, introducing
the concept of a timestamp which indicates the password age.

In this protocol, the client side only needs to know keys derived
from the timestamp and the master password, not the master password
itself.