Finish deferral of the multi-file transfers. #6885
This allows the Uploader to defer sending the multifile transfer
upload summaries, allowing a single invocation of the plugin to be
used for multiple file transfers.

In order to keep wire protocol capability, we only send updates if file
transfer object is in asynchronous mode; otherwise, we do files one-at-a-time.

Tested with the single and multiple input / output URLs cases, as well as simple
CEDAR cases.