Rate-limit the number of CCB requests in flight per CCB.
This creates a crude queue per-CCB address that limits the number
of pending messages to the remote CCB.  Protects against overwhelming
any single remote CCB server.

This could be significantly improved by sending the messages as
soon as connect() completes; the socket backlog is the scarce resource
we are currently avoiding overwhelming.

We could do even more re-engineering and reuse the same TCP socket
each time and avoid socket issues.